fix(仪表板): 日期过滤组件日期范围参数化切换至普通字段时动态时间显示异常
