+++ core/core-frontend/src/custom-component/v-query/options.ts
@@ -77,4 +77,63 @@ const addQueryCriteriaConfig = () => {
   return infoFormat(componentInfo)
 }
 
-export { infoFormat, addQueryCriteriaConfig }
+const attributes = {
+  timeType: 'fixed',
+  hideConditionSwitching: false,
+  required: false,
+  defaultMapValue: [],
+  mapValue: [],
+  parametersStart: null,
+  conditionType: 0,
+  conditionValueOperatorF: 'eq',
+  conditionValueF: '',
+  conditionValueOperatorS: 'like',
+  conditionValueS: '',
+  resultMode: 0,
+  defaultConditionValueOperatorF: 'eq',
+  defaultConditionValueF: '',
+  defaultConditionValueOperatorS: 'like',
+  defaultConditionValueS: '',
+  parametersEnd: null,
+  relativeToCurrent: 'custom',
+  timeNum: 0,
+  relativeToCurrentRange: 'custom',
+  relativeToCurrentType: 'year',
+  around: 'f',
+  arbitraryTime: new Date(),
+  timeNumRange: 0,
+  relativeToCurrentTypeRange: 'year',
+  aroundRange: 'f',
+  treeDatasetId: '',
+  displayId: '',
+  sortId: '',
+  sort: 'asc',
+  arbitraryTimeRange: new Date(),
+  setTimeRange: false,
+  showEmpty: false,
+  defaultNumValueStart: null,
+  defaultNumValueEnd: null,
+  numValueEnd: null,
+  numValueStart: null,
+  displayFormat: 0,
+  timeRange: {
+    intervalType: 'none',
+    dynamicWindow: false,
+    maximumSingleQuery: 0,
+    regularOrTrends: 'fixed',
+    regularOrTrendsValue: '',
+    relativeToCurrent: 'custom',
+    timeNum: 0,
+    relativeToCurrentType: 'year',
+    around: 'f',
+    timeNumRange: 0,
+    relativeToCurrentTypeRange: 'year',
+    aroundRange: 'f'
+  },
+  oldTreeLoad: false,
+  treeCheckedList: [],
+  defaultValueFirstItem: false,
+  treeFieldList: []
+}
+
+export { infoFormat, addQueryCriteriaConfig, attributes }
